proto: collect client tests into separate file
